+++ common/client-libs/validator-client/src/nymd/mod.rs
@@ -18,16 +18,18 @@ use execute::execute;
 pub use fee::gas_price::GasPrice;
 use mixnet_contract_common::mixnode::DelegationEvent;
 use mixnet_contract_common::{
-    ContractStateParams, Delegation, ExecuteMsg, Gateway, GatewayBond, GatewayOwnershipResponse,
-    IdentityKey, Interval, LayerDistribution, MixNode, MixNodeBond, MixOwnershipResponse,
-    MixnetContractVersion, MixnodeRewardingStatusResponse, PagedDelegatorDelegationsResponse,
-    PagedGatewayResponse, PagedMixDelegationsResponse, PagedMixnodeResponse,
-    PagedRewardedSetResponse, QueryMsg, RewardedSetUpdateDetails,
+    ContractStateParams, Delegation, ExecuteMsg, Gateway, GatewayBond, GatewayBondResponse,
+    GatewayOwnershipResponse, IdentityKey, Interval, LayerDistribution, MixNode, MixNodeBond,
+    MixOwnershipResponse, MixnetContractVersion, MixnodeBondResponse,
+    MixnodeRewardingStatusResponse, PagedDelegatorDelegationsResponse, PagedGatewayResponse,
+    PagedMixDelegationsResponse, PagedMixnodeResponse, PagedRewardedSetResponse, QueryMsg,
+    RewardedSetUpdateDetails,
 };
 use network_defaults::DEFAULT_NETWORK;
 use serde::Serialize;
 use std::convert::TryInto;
 use vesting_contract_common::ExecuteMsg as VestingExecuteMsg;
+use vesting_contract_common::QueryMsg as VestingQueryMsg;
 
 pub use crate::nymd::cosmwasm_client::client::CosmWasmClient;
 pub use crate::nymd::cosmwasm_client::signing_client::SigningCosmWasmClient;
@@ -380,6 +382,16 @@ impl<C> NymdClient<C> {
             .await
     }
 
+    pub async fn vesting_get_locked_pledge_cap(&self) -> Result<Uint128, NymdError>
+    where
+        C: CosmWasmClient + Sync,
+    {
+        let request = VestingQueryMsg::GetLockedPledgeCap {};
+        self.client
+            .query_contract_smart(self.vesting_contract_address(), &request)
+            .await
+    }
+
     pub async fn get_delegator_rewards(
         &self,
         address: String,
@@ -421,6 +433,16 @@ impl<C> NymdClient<C> {
         C: CosmWasmClient + Sync,
     {
         let request = QueryMsg::GetCurrentEpoch {};
+        self.client
+            .query_contract_smart(self.mixnet_contract_address(), &request)
+            .await
+    }
+
+    pub async fn get_current_operator_cost(&self) -> Result<u64, NymdError>
+    where
+        C: CosmWasmClient + Sync,
+    {
+        let request = QueryMsg::GetCurrentOperatorCost {};
         self.client
             .query_contract_smart(self.mixnet_contract_address(), &request)
             .await
@@ -589,6 +611,38 @@ impl<C> NymdClient<C> {
             address: address.to_string(),
         };
         let response: GatewayOwnershipResponse = self
+            .client
+            .query_contract_smart(self.mixnet_contract_address(), &request)
+            .await?;
+        Ok(response.gateway)
+    }
+
+    /// Checks whether there is a bonded mixnode associated with the provided identity key
+    pub async fn get_mixnode_bond(
+        &self,
+        identity: IdentityKey,
+    ) -> Result<Option<MixNodeBond>, NymdError>
+    where
+        C: CosmWasmClient + Sync,
+    {
+        let request = QueryMsg::GetMixnodeBond { identity };
+        let response: MixnodeBondResponse = self
+            .client
+            .query_contract_smart(self.mixnet_contract_address(), &request)
+            .await?;
+        Ok(response.mixnode)
+    }
+
+    /// Checks whether there is a bonded gateway associated with the provided identity key
+    pub async fn get_gateway_bond(
+        &self,
+        identity: IdentityKey,
+    ) -> Result<Option<GatewayBond>, NymdError>
+    where
+        C: CosmWasmClient + Sync,
+    {
+        let request = QueryMsg::GetGatewayBond { identity };
+        let response: GatewayBondResponse = self
             .client
             .query_contract_smart(self.mixnet_contract_address(), &request)
             .await?;
@@ -947,6 +1001,18 @@ impl<C> NymdClient<C> {
             VestingExecuteMsg::ClaimDelegatorReward { mix_identity },
             fee,
         )
+    }
+
+    #[execute("vesting")]
+    fn _vesting_update_locked_pledge_cap(
+        &self,
+        amount: Uint128,
+        fee: Option<Fee>,
+    ) -> (VestingExecuteMsg, Option<Fee>)
+    where
+        C: SigningCosmWasmClient + Sync,
+    {
+        (VestingExecuteMsg::UpdateLockedPledgeCap { amount }, fee)
     }
 
     /// Announce a mixnode, paying a fee.