+++ common/cosmwasm-smart-contracts/mixnet-contract/src/msg.rs
@@ -115,6 +115,7 @@ pub enum ExecuteMsg {
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum QueryMsg {
+    GetCurrentOperatorCost {},
     GetRewardingValidatorAddress {},
     GetAllDelegationKeys {},
     DebugGetAllDelegationValues {},
@@ -132,6 +133,12 @@ pub enum QueryMsg {
     },
     OwnsGateway {
         address: String,
+    },
+    GetMixnodeBond {
+        identity: IdentityKey,
+    },
+    GetGatewayBond {
+        identity: IdentityKey,
     },
     StateParams {},
     // gets all [paged] delegations associated with particular mixnode
@@ -159,6 +166,7 @@ pub enum QueryMsg {
     LayerDistribution {},
     GetRewardPool {},
     GetCirculatingSupply {},
+    GetStakingSupply {},
     GetIntervalRewardPercent {},
     GetSybilResistancePercent {},
     GetActiveSetWorkFactor {},