+++ contracts/vesting/src/contract.rs
@@ -1,5 +1,8 @@
 use crate::errors::ContractError;
-use crate::storage::{account_from_address, ADMIN, MIXNET_CONTRACT_ADDRESS};
+use crate::storage::{
+    account_from_address, locked_pledge_cap, update_locked_pledge_cap, ADMIN,
+    MIXNET_CONTRACT_ADDRESS,
+};
 use crate::traits::{
     DelegatingAccount, GatewayBondingAccount, MixnodeBondingAccount, VestingAccount,
 };
@@ -7,7 +10,7 @@ use crate::vesting::{populate_vesting_periods, Account};
 use config::defaults::DENOM;
 use cosmwasm_std::{
     coin, entry_point, to_binary, BankMsg, Coin, Deps, DepsMut, Env, MessageInfo, QueryResponse,
-    Response, Timestamp,
+    Response, Timestamp, Uint128,
 };
 use mixnet_contract_common::{Gateway, IdentityKey, MixNode};
 use vesting_contract_common::events::{
@@ -20,6 +23,8 @@ use vesting_contract_common::messages::{
     ExecuteMsg, InitMsg, MigrateMsg, QueryMsg, VestingSpecification,
 };
 use vesting_contract_common::{OriginalVestingResponse, Period, PledgeData};
+
+pub const INITIAL_LOCKED_PLEDGE_CAP: Uint128 = Uint128::new(100_000_000_000);
 
 #[entry_point]
 pub fn instantiate(
@@ -47,6 +52,9 @@ pub fn execute(
     msg: ExecuteMsg,
 ) -> Result<Response, ContractError> {
     match msg {
+        ExecuteMsg::UpdateLockedPledgeCap { amount } => {
+            try_update_locked_pledge_cap(amount, info, deps)
+        }
         ExecuteMsg::TrackReward { amount, address } => {
             try_track_reward(deps, info, amount, &address)
         }
@@ -116,6 +124,18 @@ pub fn execute(
             try_update_staking_address(to_address, info, deps)
         }
     }
+}
+
+pub fn try_update_locked_pledge_cap(
+    amount: Uint128,
+    info: MessageInfo,
+    deps: DepsMut,
+) -> Result<Response, ContractError> {
+    if info.sender != ADMIN.load(deps.storage)? {
+        return Err(ContractError::NotAdmin(info.sender.as_str().to_string()));
+    }
+    update_locked_pledge_cap(amount, deps.storage)?;
+    Ok(Response::default())
 }
 
 pub fn try_update_mixnode_config(
@@ -421,6 +441,7 @@ fn try_create_periodic_vesting_account(
 #[entry_point]
 pub fn query(deps: Deps<'_>, env: Env, msg: QueryMsg) -> Result<QueryResponse, ContractError> {
     let query_res = match msg {
+        QueryMsg::GetLockedPledgeCap {} => to_binary(&get_locked_pledge_cap(deps)),
         QueryMsg::LockedCoins {
             vesting_account_address,
             block_time,
@@ -493,6 +514,10 @@ pub fn query(deps: Deps<'_>, env: Env, msg: QueryMsg) -> Result<QueryResponse, C
     };
 
     Ok(query_res?)
+}
+
+pub fn get_locked_pledge_cap(deps: Deps<'_>) -> Uint128 {
+    locked_pledge_cap(deps.storage)
 }
 
 pub fn try_get_current_vesting_period(