+++ contracts/vesting/src/vesting/account/delegating_account.rs
@@ -1,7 +1,9 @@
 use crate::errors::ContractError;
+use crate::storage::locked_pledge_cap;
 use crate::storage::save_delegation;
 use crate::storage::MIXNET_CONTRACT_ADDRESS;
 use crate::traits::DelegatingAccount;
+use crate::traits::VestingAccount;
 use cosmwasm_std::{wasm_execute, Coin, Env, Response, Storage, Uint128};
 use mixnet_contract_common::ExecuteMsg as MixnetExecuteMsg;
 use mixnet_contract_common::IdentityKey;
@@ -58,6 +60,15 @@ impl DelegatingAccount for Account {
         storage: &mut dyn Storage,
     ) -> Result<Response, ContractError> {
         let current_balance = self.load_balance(storage)?;
+        let total_pledged_after = self.total_pledged_locked(storage, env)? + coin.amount;
+        let locked_pledge_cap = locked_pledge_cap(storage);
+
+        if locked_pledge_cap < total_pledged_after {
+            return Err(ContractError::LockedPledgeCapReached {
+                current: total_pledged_after,
+                cap: locked_pledge_cap,
+            });
+        }
 
         if current_balance < coin.amount {
             return Err(ContractError::InsufficientBalance(