+++ explorer/src/pages/Mixnodes/index.tsx
@@ -245,9 +245,12 @@ export const PageMixnodes: React.FC = () => {
     },
     {
       field: 'avg_uptime',
-      headerName: 'Avg. Uptime',
+      headerName: 'Routing Score',
       renderHeader: () => (
-        <CustomColumnHeading headingTitle="Avg. Uptime" tooltipInfo="Node’s average uptime in the last 24h." />
+        <CustomColumnHeading
+          headingTitle="Routing Score"
+          tooltipInfo="Node’s routing score is relative to that of the network. Each time a node is tested, the test packets have to go through the full path of the network (a gateway + 3 nodes). If a node in the path drop packets it will affect the score of other nodes in the test."
+        />
       ),
       headerClassName: 'MuiDataGrid-header-override',
       width: 160,